+++ app/src/main/java/com/paw/key/data/dto/request/LoginRequestDto.kt
@@ -9,3 +9,9 @@ data class LoginRequestDto (
     val email: String,
 )
 // 테스트용입니다
+
+
+fun LoginRequestDto.toEntity(): LoginRequestDto {
+    val email = this.email
+    return LoginRequestDto(email)
+}
