+++ app/src/main/java/com/paw/key/data/remote/datasource/login/AuthRemoteDataSource.kt
@@ -1,0 +1,8 @@
+package com.paw.key.data.remote.datasource.login
+
+import com.paw.key.data.dto.response.BaseResponse
+import com.paw.key.data.dto.response.LoginResponseDto
+
+interface AuthRemoteDataSource {
+    suspend fun login(providerToken: String, provider: String): BaseResponse<LoginResponseDto>
+}
