+++ app/src/main/java/com/paw/key/data/repositoryimpl/login/AuthRepositoryImpl.kt
@@ -1,7 +1,7 @@
 package com.paw.key.data.repositoryimpl.login
 
 import android.content.Context
-import com.paw.key.core.util.PreferenceDataStore
+import com.paw.key.core.util.UserDataStore
 import com.paw.key.core.util.suspendRunCatching
 import com.paw.key.data.dto.response.LoginResponseDto
 import com.paw.key.data.remote.datasource.login.AuthRemoteDataSource
@@ -12,16 +12,23 @@ import javax.inject.Inject
 class AuthRepositoryImpl @Inject constructor(
     private val authRemoteDataSource: AuthRemoteDataSource,
     private val googleAuthDataSource: GoogleAuthDataSource,
+    private val context: Context
 ) : AuthRepository {
+
     override suspend fun signInWithGoogle(context: Context): Result<String> =
         googleAuthDataSource.signIn(context).map { it.idToken }
 
     override suspend fun login(providerToken: String, provider: String): Result<LoginResponseDto> =
         suspendRunCatching {
             val loginResponse = authRemoteDataSource.login(providerToken, provider).data
-            PreferenceDataStore.saveTokens(
-                accessToken = loginResponse.AccessToken,
-                refreshToken = loginResponse.RefreshToken
+
+            UserDataStore.saveAcessToken(
+                context = context,
+                token = loginResponse.AccessToken
+            )
+            UserDataStore.saveRefreshToken(
+                context = context,
+                token = loginResponse.RefreshToken
             )
             loginResponse
         }
