+++ app/src/main/java/com/paw/key/presentation/ui/login/state/LoginContract.kt
@@ -2,23 +2,21 @@ package com.paw.key.presentation.ui.login.state
 
 import androidx.compose.runtime.Immutable
 
-class LoginContract {
-    @Immutable
-    data class LoginState(
-        val email: String = "",
-        val password: String = "",
+@Immutable
+data class LoginState(
+    val email: String = "",
+    val password: String = "",
 
-        val isPasswordVisible: Boolean = false
-    ) {
-        val isLoginValid get() = email.isNotBlank() && password.isNotBlank()
-    }
+    val isPasswordVisible: Boolean = false,
+) {
+    val isLoginValid get() = email.isNotBlank() && password.isNotBlank()
+}
 
-    sealed class LoginSideEffect {
-        data class ShowSnackBar(val message: String) : LoginSideEffect()
-        data object NavigateUp: LoginSideEffect()
-        data object NavigateNext: LoginSideEffect()
+sealed class LoginSideEffect {
+    data class ShowSnackBar(val message: String) : LoginSideEffect()
+    data object NavigateUp : LoginSideEffect()
+    data object NavigateNext : LoginSideEffect()
 
-        data object SignInSucceed : LoginSideEffect()
-        data object SignInFailed : LoginSideEffect()
-    }
+    data object SignInSucceed : LoginSideEffect()
+    data object SignInFailed : LoginSideEffect()
 }
